Small fixes for temporary directory change, and others.
The previous hard coded working_directory variable had a trailing folder
separator. But the directory name generated by TemporaryDirectory does
not have a separator, leading to the script failing.

While here I also fixed the name of the file, as I'd forgotten to add
the .csv file extension, and I corrected the mode used when opening the
csv file, so it's writeable.