Change diffing_drop_status parameter
Rachelle Ferrer suggested the diffing_drop_status parameter be changed
from inactive to deleted_last_completed. She explained that this option
is not technically a status, but combines the deleted and completed
statuses. If there is at least one other active enrollment in the
course, enrollments marked as deleted_last_completed will be deleted.
If it is the last enrollment in the course, the deleted_last_completed
enrollment will be set as complete.

Rachelle Ferrer also recommended diffing is only done for enrolments,
but this script has not yet been modified to account for that
recommendation.